requirements.txt
flask
PyPDF2
python-docx
requests
dotenv